CDPS-1858: Refactor 13/14 - globalNumbersController.ts
